+++ src/snyk/common/constants/globalState.ts
@@ -1,3 +1,4 @@
 export const MEMENTO_ANONYMOUS_ID = 'snyk.anonymousId';
 export const MEMENTO_CLI_VERSION = 'snyk.cliVersion';
 export const MEMENTO_CLI_CHECKSUM = 'snyk.cliChecksum';
+export const MEMENTO_ANALYTICS_PLUGIN_INSTALLED_SENT = 'snyk.pluginInstalledSent';