+++ src/test/unit/common/languageServer/languageServer.test.ts
@@ -248,7 +248,7 @@ suite('Language Server', () => {
         folderConfigs: [],
         authenticationMethod: 'oauth',
         enableSnykOSSQuickFixCodeActions: 'false',
-        hoverVerbosity: 0,
+        hoverVerbosity: 1,
       };
 
       deepStrictEqual(await languageServer.getInitializationOptions(), expectedInitializationOptions);